Revert "Improve Unittests (#56)"
This reverts commit 66ed177bdd82e4388ca86d94b71cf7ce349314dd.
