--- src/oehen.arguard.test/ArgumentLessOrEqualThanZeroGuardTest.cs
@@ -1,51 +1,0 @@
-using System;
-using FluentAssertions;
-using Xunit;
-
-namespace oehen.arguard
-{
-    public class ArgumentLessOrEqualThanZeroGuardTest
-    {
-        [Theory]
-        [InlineData(-1)]
-        [InlineData(-2)]
-        [InlineData(-3)]
-        [InlineData(-123456)]
-        public void ThrowIfIsLessOrEqualThanZero_ShouldThrowArgumentOutOfRangeException_IfArgumentValueIsNegative(int argument)
-        {
-            var exception = Assert.Throws<ArgumentOutOfRangeException>(() =>
-            {
-                argument.ThrowIfIsLessOrEqualThanZero(nameof(argument));
-            });
-            exception.Message.Should().Be("Argument is less or equal than 0. (Parameter 'argument')" +
-                                          Environment.NewLine + $"Actual value was {argument}.");
-        }
-
-        [Fact]
-        public void ThrowIfIsLessOrEqualThanZero_ShouldThrowArgumentOutOfRangeException_IfArgumentIsZero()
-        {
-            const int argument = 0;
-            var exception = Assert.Throws<ArgumentOutOfRangeException>(() =>
-            {
-                argument.ThrowIfIsLessOrEqualThanZero(nameof(argument));
-            });
-            exception.Message.Should().Be("Argument is less or equal than 0. (Parameter 'argument')" +
-                                          Environment.NewLine + "Actual value was 0.");
-        }
-
-        [Theory]
-        [InlineData(1)]
-        [InlineData(2)]
-        [InlineData(3)]
-        [InlineData(123456)]
-        [InlineData(int.MaxValue)]
-        public void ThrowIfIsLessOrEqualThanZero_ShouldNotThrowArgumentOutOfRangeException_IfArgumentIsMoreThanZero(int argument)
-        {
-            var exception = Record.Exception(() =>
-            {
-                argument.ThrowIfIsLessOrEqualThanZero(nameof(argument));
-            });
-            Assert.Null(exception);
-        }
-    }
-}
