+++ src/oehen.arguard.test/ArgumentLessThanGuardTest.cs
@@ -7,57 +7,84 @@ namespace oehen.arguard
     public class ArgumentLessThanGuardTest
     {
         [Theory]
-        [InlineData(-1, 0)]
-        [InlineData(2, 3)]
-        [InlineData(-3, 5)]
-        [InlineData(5, 5)]
-        public void ThrowIfIsLessOrEqualThanCompareValue_ShouldThrowArgumentOutOfRangeException_IfArgumentValueIsLessOrEqualThanTheCompareValue(int argument, int compareValue)
+        [InlineData(-1)]
+        [InlineData(-2)]
+        [InlineData(-3)]
+        [InlineData(-123456)]
+        public void ThrowIfIsLessOrEqualThanZero_ShouldThrowArgumentOutOfRangeException_IfArgumentValueIsNegative(int argument)
         {
             var exception = Assert.Throws<ArgumentOutOfRangeException>(() =>
             {
-                argument.ThrowIfIsLessOrEqualThan(compareValue, nameof(argument));
+                argument.ThrowIfIsLessOrEqualThanZero(nameof(argument));
             });
-            exception.Message.Should().Be($"Argument is less or equal than {compareValue}. (Parameter 'argument')" +
+            exception.Message.Should().Be("Argument is less or equal than 0. (Parameter 'argument')" +
                                           Environment.NewLine + $"Actual value was {argument}.");
         }
 
         [Theory]
-        [InlineData(-1, 0)]
-        [InlineData(2, 3)]
-        [InlineData(-3, 5)]
-        public void ThrowIfIsLessThanCompareValue_ShouldThrowArgumentOutOfRangeException_IfArgumentValueIsLessThanTheCompareValue(int argument, int compareValue)
+        [InlineData(-1)]
+        [InlineData(-2)]
+        [InlineData(-3)]
+        [InlineData(-123456)]
+        public void ThrowIfIsLessThanZero_ShouldThrowArgumentOutOfRangeException_IfArgumentValueIsNegative(int argument)
         {
             var exception = Assert.Throws<ArgumentOutOfRangeException>(() =>
             {
-                argument.ThrowIfIsLessThan(compareValue, nameof(argument));
+                argument.ThrowIfIsLessThanZero(nameof(argument));
             });
-            exception.Message.Should().Be($"Argument is less than {compareValue}. (Parameter 'argument')" +
+            exception.Message.Should().Be("Argument is less than 0. (Parameter 'argument')" +
                                           Environment.NewLine + $"Actual value was {argument}.");
         }
 
-        [Theory]
-        [InlineData(11, 2)]
-        [InlineData(3, 2)]
-        [InlineData(5, 3)]
-        [InlineData(3, 3)]
-        public void ThrowIfIsLessThanCompareValue_ShouldNotThrowArgumentOutOfRangeException_IfArgumentValueIsGreaterThanTheCompareValue(int argument, int compareValue)
+        [Fact]
+        public void ThrowIfIsLessOrEqualThanZero_ShouldThrowArgumentOutOfRangeException_IfArgumentIsZero()
         {
+            const int argument = 0;
+            var exception = Assert.Throws<ArgumentOutOfRangeException>(() =>
+            {
+                argument.ThrowIfIsLessOrEqualThanZero(nameof(argument));
+            });
+            exception.Message.Should().Be("Argument is less or equal than 0. (Parameter 'argument')" +
+                                          Environment.NewLine + "Actual value was 0.");
+        }
+
+        [Fact]
+        public void ThrowIfIsLessThanZero_ShouldNotThrowArgumentOutOfRangeException_IfArgumentIsZero()
+        {
+            const int argument = 0;
             var exception = Record.Exception(() =>
             {
-                argument.ThrowIfIsLessThan(compareValue, nameof(argument));
+                argument.ThrowIfIsLessThanZero(nameof(argument));
             });
             Assert.Null(exception);
         }
 
         [Theory]
-        [InlineData(11, 2)]
-        [InlineData(3, 2)]
-        [InlineData(5, 3)]
-        public void ThrowIfIsLessOrEqualThanCompareValue_ShouldNotThrowArgumentOutOfRangeException_IfArgumentValueIsGreaterOrEqualThanTheCompareValue(int argument, int compareValue)
+        [InlineData(1)]
+        [InlineData(2)]
+        [InlineData(3)]
+        [InlineData(123456)]
+        [InlineData(int.MaxValue)]
+        public void ThrowIfIsLessOrEqualThanZero_ShouldNotThrowArgumentOutOfRangeException_IfArgumentIsMoreThanZero(int argument)
         {
             var exception = Record.Exception(() =>
             {
-                argument.ThrowIfIsLessOrEqualThan(compareValue, nameof(argument));
+                argument.ThrowIfIsLessOrEqualThanZero(nameof(argument));
+            });
+            Assert.Null(exception);
+        }
+
+        [Theory]
+        [InlineData(1)]
+        [InlineData(2)]
+        [InlineData(3)]
+        [InlineData(123456)]
+        [InlineData(int.MaxValue)]
+        public void ThrowIfIsLessThanZero_ShouldNotThrowArgumentOutOfRangeException_IfArgumentIsMoreThanZero(int argument)
+        {
+            var exception = Record.Exception(() =>
+            {
+                argument.ThrowIfIsLessThanZero(nameof(argument));
             });
             Assert.Null(exception);
         }
